apps/bfagent/docs_v2/bfa_agent update and thoughts/pyproject.toml
[project]
name = "bfa-agent"
version = "0.1.0"
description = "Explosionsschutz-Agent mit OpenAI Agents SDK + OpenRouter"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.1.0",
    "openai>=1.50.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
mcp = [
    "mcp>=1.0.0",
    "fastmcp>=0.4.0",
]
research = [
    "paper-search-mcp>=0.1.0",
]
full = [
    "mcp>=1.0.0",
    "fastmcp>=0.4.0",
    "paper-search-mcp>=0.1.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.5.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
